Resolve circular import
By directly calling the logger from the main simplemonitor loop we can
avoid needing to pass the Logger object into the Monitor just to have it
call the save_result2() method on it. This removes the need to import
both modules into each other in order to decorate the types, which
caused a circular dependency for imports.